requirements.txt
pydantic ~= 2.4
Pillow ~= 10.1
aiohttp ~= 3.8
